Move version function to API
